Correction to Stat type

+++ philameter/DAL/Entities/Stats.cs
@@ -11,7 +11,7 @@ namespace philameter.DAL.Entities
         public string Name { get; set; }
         public string Details { get; set; }
         public string Source { get; set; }
-        public int Stat { get; set; }       
+        public double Stat { get; set; }       
 
     }
 }
